pyproject.toml
[tool.poetry]
name = "eta-estimator"
version = "0.1.0"
description = "A Django project with PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
pytest-django = "^4.11.1"
pandas = "^2.2.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 